+++ lib/ui/screens/cart/cart_screen.dart
@@ -13,45 +13,22 @@ import 'package:bubble_tab_indicator/bubble_tab_indicator.dart';
 import 'package:w3/core/constants/buttons_style.dart';
 import 'package:w3/ui/screens/signup/signup_screen.dart';
 import 'package:w3/ui/custom_widgets/home_screen_costum_widgets.dart';
+import 'package:w3/ui/custom_widgets/my_cart_cards.dart';
 
 class CartScreen extends StatelessWidget {
   CartScreen({Key? key}) : super(key: key);
 
   List<Widget> list = [
-    MyCartList(
+    MyCartCards(
       plantDiscribText: "asd",
       plantImageUrl: "assets/images/plant1.png",
       plantNameText: "Gullab1",
       plantPrice: 12.00,
     ),
-    MyCartList(
+    MyCartCards(
       plantDiscribText: "asdfg",
       plantImageUrl: "assets/images/plant1.png",
       plantNameText: "Gullab2",
-      plantPrice: 12.00,
-    ),
-    MyCartList(
-      plantDiscribText: "khgf",
-      plantImageUrl: "assets/images/plant1.png",
-      plantNameText: "Gullab3",
-      plantPrice: 12.00,
-    ),
-    MyCartList(
-      plantDiscribText: "ljgf",
-      plantImageUrl: "assets/images/plant1.png",
-      plantNameText: "Gullab4",
-      plantPrice: 12.00,
-    ),
-    MyCartList(
-      plantDiscribText: "kytr",
-      plantImageUrl: "assets/images/plant1.png",
-      plantNameText: "Gullab5",
-      plantPrice: 12.00,
-    ),
-    MyCartList(
-      plantDiscribText: "jhfdef",
-      plantImageUrl: "assets/images/plant1.png",
-      plantNameText: "Gullab6",
       plantPrice: 12.00,
     ),
   ];
@@ -228,202 +205,6 @@ class CartScreen extends StatelessWidget {
           ),
         ]),
       ),
-    );
-  }
-}
-
-class MyCartList extends StatelessWidget {
-  MyCartList({
-    Key? key,
-    required this.plantImageUrl,
-    required this.plantNameText,
-    required this.plantDiscribText,
-    required this.plantPrice,
-    // required this.onTapFunction,
-  }) : super(key: key);
-
-  // plant image url
-  final String plantImageUrl;
-
-  // plant name text
-  final String plantNameText;
-
-  // plant type text
-  final String plantDiscribText;
-
-  // plant price
-  final double? plantPrice;
-
-  //final VoidCallback onTapFunction;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.only(bottom: 15.h),
-      child: Container(
-        height: 78.h,
-        width: MediaQuery.of(context).size.width,
-        child: Padding(
-          padding:
-              EdgeInsets.only(top: 10.h, left: 10.w, bottom: 10.h, right: 5.w),
-          child: Row(
-            children: [
-              // small plant image
-              Expanded(
-                flex: 2,
-                child: Container(
-                  width: 80.w,
-                  height: 80.h,
-                  child: Center(
-                    child: Image.asset(
-                      plantImageUrl,
-                      fit: BoxFit.contain,
-                      height: 50.h,
-                      width: 50.w,
-                    ),
-                  ),
-                  decoration: BoxDecoration(
-                    color: Colors.amber.shade100,
-                    borderRadius: BorderRadius.circular(15.0),
-                  ),
-                ),
-              ),
-              SizedBox(
-                width: 15.w,
-              ),
-
-              // plant name
-              Expanded(
-                flex: 6,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    // plant name
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Text(
-                          plantNameText,
-                          style: TextStyle(
-                            color: kDarkGreen,
-                            fontFamily: "Poppins",
-                            fontSize: 12.sp,
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-
-                        //
-                        Padding(
-                          padding: EdgeInsets.only(right: 5.w),
-                          child: GestureDetector(
-                            onTap: () {},
-                            child: const Icon(Icons.more_vert),
-                          ),
-                        ),
-                      ],
-                    ),
-
-                    // plant discription
-                    Text(
-                      plantDiscribText,
-                      style: TextStyle(
-                        color: Colors.grey,
-                        fontFamily: "Poppins",
-                        fontSize: 10.sp,
-                        fontWeight: FontWeight.w600,
-                      ),
-                    ),
-                    SizedBox(
-                      height: 10.h,
-                    ),
-
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Row(
-                          children: [
-                            AddToCartAndRemoveButton(
-                              iconData: Icons.remove,
-                              onTap: () {},
-                            ),
-
-                            Padding(
-                              padding: EdgeInsets.symmetric(horizontal: 10.w),
-                              //
-                              child: Container(
-                                width: 3.w,
-                                height: 10.h,
-                                color: kDarkGreen,
-                              ),
-                            ),
-
-                            //add to cart
-                            AddToCartAndRemoveButton(
-                              iconData: Icons.add,
-                              onTap: () {},
-                            ),
-                          ],
-                        ),
-                        Padding(
-                          padding: EdgeInsets.only(right: 10.w),
-                          child: Text(
-                            "\$$plantPrice",
-                            style:
-                                plantPriceTextStyle.copyWith(fontSize: 13.sp),
-                          ),
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-        decoration: BoxDecoration(
-          color: kLightenDarkGreenColor,
-          borderRadius: BorderRadius.circular(15.0),
-          boxShadow: [
-            BoxShadow(
-              color: kGreen.withOpacity(0.1),
-              spreadRadius: 1.r,
-              blurRadius: 1.r,
-              offset: Offset(0, 2), // changes position of shadow
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class AddToCartAndRemoveButton extends StatelessWidget {
-  const AddToCartAndRemoveButton(
-      {Key? key, required this.onTap, required this.iconData})
-      : super(key: key);
-
-  final VoidCallback onTap;
-  final IconData iconData;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-          width: 20.w,
-          height: 15.h,
-          child: Center(
-            child: Icon(
-              iconData,
-              size: 10.h,
-            ),
-          ),
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(5.r),
-            border: Border.all(
-              color: kDarkGreen,
-            ),
-          )),
     );
   }
 }
