+++ lib/ui/screens/menu/menu_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class MenuScreen extends StatelessWidget {
+  const MenuScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: Text("Profile Screen"));
+  }
+}
