+++ lib/ui/screens/root_screen.dart
@@ -6,6 +6,8 @@ import 'package:w3/ui/screens/home/home_screen.dart';
 import 'package:w3/ui/custom_widgets/custom_widgets.dart';
 import 'package:w3/core/constants/text_style.dart';
 import 'package:w3/core/constants/buttons_style.dart';
+import 'package:w3/ui/screens/menu/menu_Screen.dart';
+import 'package:w3/ui/screens/profile/profile_Screen.dart';
 import 'package:w3/ui/screens/signup/signup_screen.dart';
 
 class RootScreen extends StatefulWidget {
@@ -19,9 +21,9 @@ class _RootScreenState extends State<RootScreen> {
   int _selectedIndex = 0;
   List<Widget> pagesList = [
     HomeScreen(),
+    ProfileScreen(),
     CartScreen(),
-    CartScreen(),
-    CartScreen(),
+    MenuScreen(),
   ];
 
   //button navigation bar items ontab function
